+++ src/Products/PloneMeeting/browser/views.py
@@ -43,6 +43,7 @@ from plone.memoize.view import memoize_contextless
 from Products.CMFCore.permissions import ManagePortal
 from Products.CMFCore.permissions import ModifyPortalContent
 from Products.CMFCore.utils import _checkPermission
+from Products.CMFPlone.utils import safe_unicode
 from Products.Five import BrowserView
 from Products.PloneMeeting import logger
 from Products.PloneMeeting.browser.itemchangeorder import _is_integer
@@ -663,7 +664,7 @@ class BaseDGHV(object):
                     hasSeparation = False
                     if checkNeedSeparator:
                         preparedXhtmlContent = "<special_tag>%s</special_tag>" % xhtmlContent
-                        tree = lxml.html.fromstring(unicode(preparedXhtmlContent, 'utf-8'))
+                        tree = lxml.html.fromstring(safe_unicode(preparedXhtmlContent))
                         children = tree.getchildren()
                         if children and not children[-1].text:
                             hasSeparation = True
@@ -772,9 +773,9 @@ class BaseDGHV(object):
                 if withDelayLabel and advice['delay'] and advice['delay_label']:
                     if delayAwareMsg:
                         delayAwareMsg = "%s - %s" % (delayAwareMsg,
-                                                     unicode(advice['delay_label'], 'utf-8'))
+                                                     safe_unicode(advice['delay_label']))
                     else:
-                        delayAwareMsg = "%s" % unicode(advice['delay_label'], 'utf-8')
+                        delayAwareMsg = "%s" % safe_unicode(advice['delay_label'])
                 if delayAwareMsg:
                     delayAwareMsg = u" <i>(%s)</i>" % cgi.escape(delayAwareMsg)
                     res = res + u"<u>%s %s:</u>" % (cgi.escape(advice['name']),
@@ -803,13 +804,13 @@ class BaseDGHV(object):
                     res = res + u"<br /><u>%s :</u> <i>%s</i>" % (translate('Advice given by',
                                                                   domain='PloneMeeting',
                                                                   context=self.request),
-                                                                  cgi.escape(unicode(author, 'utf-8')), )
+                                                                  cgi.escape(safe_unicode(author)), )
 
-                    adviceComment = 'comment' in advice and self.printXhtml(adviceObj, advice['comment']) or '-'
+                    adviceComment = advice['comment'] and self.printXhtml(adviceObj, advice['comment']) or '-'
                     res = res + (u"<br /><u>%s :</u> %s<p></p>" % (translate('Advice comment',
                                                                              domain='PloneMeeting',
                                                                              context=self.request),
-                                                                   unicode(adviceComment, 'utf-8')))
+                                                                   safe_unicode(adviceComment)))
                 res += u"</p>"
         if not itemAdvicesByType:
             res += "<p class='pmAdvices'>-</p>"